backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0

# Database and ORM
psycopg2-binary==2.9.7  # PostgreSQL adapter (optional, for production)

# API and filtering
django-filter==23.3
django-cors-headers==4.3.1
coreapi==2.3.3
Pygments==2.16.1
drf-spectacular==0.27.0

# Authentication and security
djangorestframework-simplejwt==5.3.0

# Cache and Redis
redis==5.0.1
django-redis==5.4.0

# Development and testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0

# Code quality
black==23.9.1
flake8==6.1.0
mypy==1.6.1
django-stubs==4.2.6

# Production server
gunicorn==21.2.0

# Package management
setuptools==69.0.2

# Environment management
python-decouple==3.8